Fix not being able to access prop array issue
So apparently in react when you pass an array as a prop,
it is stored in the prop object. So you need to first access the prop
, then choose your array from there. Ex: props.parts as opposed to
just directly referring to it as props[0].name .That's why it
kept saying props was just [object Object]. Because it's converted
to an object first.